requirements.txt
django-crispy-forms==1.8.1
django-cleanup==4.0.0
